Relocate git-oauth + deploy route handlers to keep routes.c under the line cap
The added GitHub web-flow and deploy handlers pushed server_http_routes.c past
the 2500-line lint ceiling. Move the whole git-oauth-github cluster (device
start/poll/config + web start/callback) and the deploy handlers next to the
existing device-flow handlers in server_http_config_routes.c (de-static + declare
in server_http_internal.h; use its device_git_surface_enabled gate). The route
table and every /v1 literal stay in server_http_routes.c, so API conformance is
unchanged. No behavior change.
